pyproject.toml
[tool.pdm]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "jaaj", email = "otaviobentto@gmail.com"},
]
dependencies = [
    "django>=4.2",
    "djangorestframework>=3.14.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}
